requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.6
django-discover-runner==0.4
Sphinx


# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.0

ipython==1.2.0
